Longest contained interval using Set
**Problem**
Write a program which takes as input a set of integers represented by an array, and returns the size of a largest subset of integers in the array having the property that if two integers are in the subset, then so are all integers between them. For example, if the input is (3,-2,7,9,8,1,2,0,-1,5,8), the largest such subset is{-2,-1,0,1,2,3), so you should return 6.